Move blender output format adjustment to queue_operations
